count stock/sell

+++ ebike-server-core/src/main/java/com/ecgobike/pojo/response/BatteryProductVO.java
@@ -10,5 +10,5 @@ public class BatteryProductVO {
     private Long productId;
     private String type;
     private String iconUrl;
-    private Integer stockNum;
+    private long stockNum;
 }
